Fix clear fault proc bug
The fault clearing logic was only clearing a fault if the subsequent
method call was on the same line as the original problem.  So, it would
clear this:

x = [1,2,3].select {|y| y > 1 } ; x.empty? ; x.first

but not this:

x = [1,2,3].select {|y| y > 1 }
x.empty?
x.first

And of course the latter was by far the more common case.