Never add the same problem twice
If we encounter something like:

x = get_numbers.select {|y| y > 1 }
if x.any?
  x.join(" ")
end

Then we'll remove it from the problem set.  However, if the next test
that runs has get_numbers set up so that x is empty, the basic block
inside the if statement will never be executed and the problem will be
readded.  Thus, once a problem is added and then removed, we should
never readd it.